refactor: move thunks to actions file
